spdmlib: fix cargo clippy warnings
- Allow clippy::duplicate_mod for intentional shared test module pattern
- Use struct initializers instead of field reassignment after Default::default()
- Derive Default for SpdmChunkStatus with #[default] attribute
- Replace if is_some()/unwrap() with if let Some()
- Remove unnecessary &mut references and casts
- Use .is_multiple_of() instead of manual modulo checks
- Replace assert_eq!(x, true/false) with assert!(x)/assert!(!x)
- Remove assert!(true/false), replace with panic!() or remove
- Use iterators instead of needless range loops
- Eliminate single-element loops
- Use unwrap_or_else instead of expect with format!
- Prefix unused variable with underscore
- Simplify boolean expressions

+++ spdmlib/src/message/challenge_test.rs
@@ -55,7 +55,7 @@ fn test_challenge_struct() {
         + INVALID_OPAQUE_DATA_LENGTH as usize
         + SHA256_DIGEST_SIZE * 2];
     LittleEndian::write_u16(
-        &mut u8_slice[(36 + 2 * SHA256_DIGEST_SIZE as usize)..],
+        &mut u8_slice[(36 + 2 * SHA256_DIGEST_SIZE)..],
         INVALID_OPAQUE_DATA_LENGTH,
     );
     let reader = &mut Reader::init(&u8_slice[2..]);
@@ -80,7 +80,7 @@ fn test_challenge_struct_opaque_data_length_negative() {
         + SHA256_DIGEST_SIZE * 2];
     const OPAQUE_DATA_LENGTH: u16 = 1024u16;
     LittleEndian::write_u16(
-        &mut u8_slice[(36 + 2 * SHA256_DIGEST_SIZE as usize)..],
+        &mut u8_slice[(36 + 2 * SHA256_DIGEST_SIZE)..],
         OPAQUE_DATA_LENGTH,
     );
     let reader = &mut Reader::init(&u8_slice[2..]);
